task_2/requirements.txt
kagglehub
seaborn
kaggle
numpy
pandas
matplotlib
Pillow
torch
torchvision
scikit-learn
transformers
datasets
evaluate
seqeval
faker
'accelerate>=1.1.0'